updated deleted action

+++ Backend/src/main/java/com/developer/Student_Management/service/implementation/StudentServiceImplementation.java
@@ -1,7 +1,7 @@
 package com.developer.Student_Management.service.implementation;
 
 import com.developer.Student_Management.dto.StudentDto;
-import com.developer.Student_Management.entity.Student;
+import com.developer.Student_Management.entity.StudentEntity;
 import com.developer.Student_Management.mapper.StudentMapper;
 import com.developer.Student_Management.repository.StudentRepository;
 import com.developer.Student_Management.service.StudentService;
@@ -18,47 +18,49 @@ public class StudentServiceImplementation implements StudentService {
     private StudentRepository studentRepository;
     @Override
     public StudentDto createStudent(StudentDto studentDto) {
-        Student student = StudentMapper.mapToStudent(studentDto);
-        Student savedStudent = studentRepository.save(student);
+        StudentEntity studentEntity = StudentMapper.mapToStudent(studentDto);
+        StudentEntity savedStudent = studentRepository.save(studentEntity);
         return StudentMapper.mapToStudentDto(savedStudent);
     }
 
     @Override
     public StudentDto getStudentByID(Long studentId) {
-        Student student = studentRepository.findById(studentId)
+        StudentEntity student = studentRepository.findById(studentId)
                 .orElseThrow(null);
         return StudentMapper.mapToStudentDto(student);
     }
 
     @Override
     public List<StudentDto> getAllStudents() {
-        List<Student> students = studentRepository.findAll();
+        List<StudentEntity> students = studentRepository.findAll();
         return students.stream().map((student )-> StudentMapper.mapToStudentDto(student))
                 .collect(Collectors.toList());
     }
 
     @Override
     public StudentDto updateStudent(Long studentId, StudentDto updatedStudent) {
-        Student student = studentRepository.findById(studentId).orElseThrow(
+        StudentEntity studentEntity = studentRepository.findById(studentId).orElseThrow(
                 () -> new RuntimeException("Student not exists with id: " + studentId));
 
-        student.setName(updatedStudent.getName());
-        student.setAge(updatedStudent.getAge());
-        student.setGender(updatedStudent.getGender());
-        student.setRollNo(updatedStudent.getRollNo());
-        student.setCourse(updatedStudent.getCourse());
-        student.setStream(updatedStudent.getStream());
-        student.setSemester(updatedStudent.getSemester());
-        student.setStudentPhoto(updatedStudent.getStudentPhoto());
-        Student updatedStudentData = studentRepository.save(student);
+        studentEntity.setName(updatedStudent.getName());
+        studentEntity.setAge(updatedStudent.getAge());
+        studentEntity.setGender(updatedStudent.getGender());
+        studentEntity.setRollNo(updatedStudent.getRollNo());
+        studentEntity.setCourse(updatedStudent.getCourse());
+        studentEntity.setStream(updatedStudent.getStream());
+        studentEntity.setSemester(updatedStudent.getSemester());
+        studentEntity.setStudentPhoto(updatedStudent.getStudentPhoto());
+        StudentEntity updatedStudentData = studentRepository.save(studentEntity);
         return StudentMapper.mapToStudentDto(updatedStudentData);
     }
 
     @Override
     public void deleteStudent(Long studentId) {
-        Student student = studentRepository.findById(studentId).orElseThrow(
-                () -> new RuntimeException("Student not exists with id: " + studentId)
-        );
+        // Check if the student exists
+        if (!studentRepository.existsById(studentId)) {
+            throw new RuntimeException("Student not exists with id: " + studentId);
+        }
+        // Delete the student
         studentRepository.deleteById(studentId);
     }
 }
